remove `self` nodelocal examples
We do not recommend `self` in most cases and showing it as an example gives it more exposure than we really desire.

We especially want to be sure that anyone using "self" has read and understood the footnote under "considerations". This note was previously linked after the example but that relied on one clicking it and reading it, so a user could presumably discover and start using `self` without reading the note.

After this change, the only mention of `self` is in the footnote, so only users who read the note should be aware it is an option, specifically in the configuration where it is a valid option.